backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
python-multipart==0.0.12
python-docx==1.1.2
google-generativeai==0.8.3
groq==1.2.0
httpx==0.28.0
python-dotenv==1.0.1
pydantic-settings==2.6.1
pydantic[email]==2.10.3
